+++ src/app/ui/(main)/plans/page.tsx
@@ -8,6 +8,7 @@ import {PAYPAL_CONFIG} from "@/lib/paypal-client";
 import {PayPalProvider} from "@paypal/react-paypal-js/sdk-v6";
 import { ROUTES } from "@/lib/routes";
 import { buildPageMetadata, normalizeLocale } from "@/lib/seo/metadata";
+import {getUserPlan, isLoggedIn} from "@/data/dto/user-dto";
 
 const META = {
   en: {
@@ -170,9 +171,12 @@ const content = {
   },
 } as const;
 
-export default function PlansPage() {
-  const locale = useLocale() as "es" | "en";
+export default async function PlansPage() {
+  const locale = await getLocale() as "es" | "en"; //useLocale() as "es" | "en";
   const t = content[locale] ?? content.en;
+
+  const logged = await isLoggedIn();
+  const plan = await getUserPlan();
 
   return (
     <PayPalProvider {...PAYPAL_CONFIG}>
@@ -189,7 +193,7 @@ export default function PlansPage() {
           <p className="text-lg text-gray-600">{t.intro}</p>
         </div>
 
-        <PricingCards logged={false} />
+        <PricingCards logged={logged} plan={plan} />
 
         <section className="max-w-2xl w-full mt-14 space-y-10 text-gray-700">
           <div>
