+++ src/lib/api/errors.ts
@@ -123,3 +123,23 @@ export class ResourceActionError extends ApiError {
     );
   }
 }
+
+export class PlanNotFoundError extends ApiError {
+  constructor(message?: string|null, options?: ApiErrOptions) {
+    super(
+        ERROR.PLAN_NOT_FOUND,
+        message ?? MESSAGE.PLAN_NOT_FOUND,
+        options ?? { type: 'resource', status: 400 }
+    );
+  }
+}
+
+export class UserAlReadyHasPlanError extends ApiError {
+  constructor(message?: string|null, options?: ApiErrOptions) {
+    super(
+        ERROR.PLAN_ALREADY,
+        message ?? MESSAGE.PLAN_ALREADY,
+        options ?? { type: 'resource', status: 400 }
+    );
+  }
+}
